Fix: Reset the project for a clean start
