+++ app-example/(tabs)/explore.tsx
