+++ app.json
@@ -1,11 +1,11 @@
 {
   "expo": {
-    "name": "MyFit-Frontend",
-    "slug": "MyFit-Frontend",
+    "name": "MyFit",
+    "slug": "myfit",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/images/icon.png",
-    "scheme": "myapp",
+    "scheme": "myfit",
     "userInterfaceStyle": "automatic",
     "splash": {
       "image": "./assets/images/splash.png",
